+++ ChangeLog.markdown
@@ -5,6 +5,7 @@ ImboClient-0.3.0
 ----------------
 __N/A__
 
+* Fixed issue #19: Method for adding images via URLs
 * Fixed issue #17: Allow passing cURL-driver options related to SSL
 * Pull request #18: Added maxSize-transformation to clients ImageUrl-interface (Espen Hovlandsdal)
 * Fixed issue #16: ImboClient does not remove an explicitly set port 443 from the URL when using https://